+++ pkg/randomgraph/random_graph_test.go
@@ -2,7 +2,7 @@ package randomgraph
 
 import (
 	"encoding/json"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/models"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/models"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
