+++ pkg/scenarioorchestrator/docker/scenario_orchestrator_test.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	dockercontainer "github.com/docker/docker/api/types/container"
 	"github.com/krkn-chaos/krknctl/pkg/config"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/models"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/test"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/models"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/test"
 	"github.com/stretchr/testify/assert"
 	"os"
 	"regexp"
