+++ pkg/scenarioorchestrator/factory/factory.go
@@ -2,11 +2,11 @@ package factory
 
 import (
 	"github.com/krkn-chaos/krknctl/pkg/config"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/docker"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/models"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/podman"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/utils"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/docker"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/models"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/podman"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/utils"
 )
 
 type ScenarioOrchestratorFactory struct {
@@ -19,7 +19,7 @@ func NewScenarioOrchestratorFactory(config config.Config) *ScenarioOrchestratorF
 	}
 }
 
-func (f *ScenarioOrchestratorFactory) NewInstance(containerEnvironment models.ContainerRuntime) scenario_orchestrator.ScenarioOrchestrator {
+func (f *ScenarioOrchestratorFactory) NewInstance(containerEnvironment models.ContainerRuntime) scenarioorchestrator.ScenarioOrchestrator {
 	switch containerEnvironment {
 	case models.Podman:
 		return f.getOrchestratorInstance(models.Podman)
@@ -32,7 +32,7 @@ func (f *ScenarioOrchestratorFactory) NewInstance(containerEnvironment models.Co
 	return nil
 }
 
-func (f *ScenarioOrchestratorFactory) getOrchestratorInstance(containerEnvironment models.ContainerRuntime) scenario_orchestrator.ScenarioOrchestrator {
+func (f *ScenarioOrchestratorFactory) getOrchestratorInstance(containerEnvironment models.ContainerRuntime) scenarioorchestrator.ScenarioOrchestrator {
 	if containerEnvironment == models.Podman {
 		return &podman.ScenarioOrchestrator{
 			Config:           f.Config,
