+++ pkg/scenarioorchestrator/podman/scenario_orchestrator_test.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"github.com/containers/podman/v5/pkg/bindings/containers"
 	"github.com/krkn-chaos/krknctl/pkg/config"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/models"
-	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/test"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/models"
+	"github.com/krkn-chaos/krknctl/pkg/scenarioorchestrator/test"
 	"github.com/stretchr/testify/assert"
 	operating_system "os"
 	"regexp"
