+++ pkg/scenarioorchestrator/utils/errors.go
